feat: add webpack; add husky & lint-staged to precommit; add mocha & tests
