+++ package.json
@@ -12,9 +12,11 @@
     "npm": ">=6.14"
   },
   "scripts": {
-    "start": "rimraf .parcel-cache && npm run build && node scripts/server.js",
-    "dev": "parcel src/index.html --port 1234 --open --no-cache",
-    "build": "parcel build src/index.html --no-cache",
+    "start": "rimraf dist && npm run build && node server.js",
+    "dev_": "parcel src/index.html --port 1234 --open --no-cache",
+    "dev": "webpack serve --progress --mode=development ",
+    "build_": "parcel build src/index.html --no-cache",
+    "build": "webpack --mode production",
     "test": "mocha -r ts-node/register -r tsconfig-paths/register -r ./test/setup.js './src/**/*.test.ts'",
     "postinstall": "husky install",
     "pre-commit": "npm run lint-staged",
@@ -44,6 +46,7 @@
     "sanitize-html": "~2.10.0"
   },
   "devDependencies": {
+    "@svgr/webpack": "^8.0.1",
     "@types/chai": "^4.3.5",
     "@types/chai-dom": "^1.11.0",
     "@types/mocha": "^10.0.1",
@@ -74,8 +77,6 @@
     "lint-staged": "^13.2.2",
     "mini-css-extract-plugin": "^2.7.6",
     "mocha": "^10.2.0",
-    "parcel": "~2.8.3",
-    "parcel-resolver-ts-base-url": "~1.3.1",
     "path-browserify": "^1.0.1",
     "postcss": "~8.4.21",
     "postcss-loader": "^7.3.3",
