--- scripts/server.js
@@ -1,14 +1,0 @@
-const express = require('express');
-
-const app = express();
-const PORT = process.env.PORT || 3000;
-
-app.use(express.static('./dist'));
-
-app.get('*', function(req, res){
-    res.sendFile(__dirname + '/dist/index.html');
-});
-
-app.listen(PORT, function() {
-    console.log(`Listening on port: ${PORT}`);
-});
